+++ R/RcppExports.R
@@ -1,6 +1,10 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+logit_glmnet <- function(X, y, alpha = 0, lambda = as.numeric( c(0))) {
+    .Call(`_daprcpp_logit_glmnet`, X, y, alpha, lambda)
+}
+
 make_matrix <- function(p, anno_df) {
     .Call(`_daprcpp_make_matrix`, p, anno_df)
 }
@@ -27,9 +31,5 @@ logit_torus <- function(X, y, lambdaL1 = 0, lambdaL2 = 0) {
 
 logit_qbinom <- function(X, y) {
     .Call(`_daprcpp_logit_qbinom`, X, y)
-}
-
-logit_glmnet <- function(X, y, alpha = 0, lambda = as.numeric( c(0))) {
-    .Call(`_daprcpp_logit_glmnet`, X, y, alpha, lambda)
 }
