+++ tests/testthat/test_data.R
@@ -16,10 +16,6 @@ test_that("annotation matrix is generated correctly",{
   expect_equal(annomat_l,list(annomat=matrix(c(1,0,1,1),4,1),
                               names=c("aa")))
   
-
-  library(Matrix)  
-  library(daprcpp)
-  library(testthat)
   p <- 4
   anno_df <- tibble::tibble(SNP=c(1L,3L,4L),feature=c("aa","aa","aa"))  
   annomat_l <- make_spmatrix_df(p,anno_df)
